refactor(bench): share browser probe pipeline
